% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.SIBCASA.R
\name{met2model.SIBCASA}
\alias{met2model.SIBCASA}
\title{Write SIBCASA met files}
\usage{
met2model.SIBCASA(in.path, in.prefix, outfolder, overwrite = FALSE)
}
\arguments{
\item{in.path}{path on disk where CF file lives}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written}

\item{overwrite}{logical: Replace output if it already exists?}
}
\value{
OK if everything was succesful.
}
\description{
Converts a met CF file to a sibcasa specific met file. The input
files are calld <in.path>/<in.prefix>.YYYY.cf
}
\author{
Tony Gardella
}
