% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.SIBCASA.R
\name{model2netcdf.SIBCASA}
\alias{model2netcdf.SIBCASA}
\title{Convert SIBCASA output into the NACP Intercomparison format (ALMA using netCDF)}
\usage{
model2netcdf.SIBCASA(outdir, sitelat, sitelon, start_date, end_date)
}
\arguments{
\item{outdir}{Location of model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\description{
Convert SIBCASA output into the NACP Intercomparison format (ALMA using netCDF)
}
\author{
Tony Gardella
}
