% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_restart.SIBCASA.R
\name{read_restart.SIBCASA}
\alias{read_restart.SIBCASA}
\title{Read restart template for SDA}
\usage{
read_restart.SIBCASA(outdir, runid, stop.time, settings, var.names, params)
}
\arguments{
\item{outdir}{Output directory}

\item{runid}{Run ID}

\item{stop.time}{Year that is being read}

\item{settings}{PEcAn settings object}

\item{var.names}{Variable names to be extracted}

\item{params}{Any parameters required for state calculations}
}
\value{
Forecast numeric matrix
}
\description{
Read restart files from model.
}
\author{
Alexey Shiklomanov
}
