% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sibcasa_output_vars}
\alias{sibcasa_output_vars}
\title{Output variables for SIBCASA}
\format{
data frame with 6 unnamed columns
\describe{
 \item{1}{? (t 0, t 3, etc)}
 \item{2}{short name}
 \item{3}{row number?}
 \item{4}{description}
 \item{5}{units}
 \item{6}{unused}}
}
\usage{
sibcasa_output_vars
}
\description{
TODO: Provide a complete description here of the dataset,
	including its origin and purpose.
 The column descriptions below were written by someone looking at the file
	with no insight into its usage
}
\keyword{datasets}
