% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.SIBCASA.R
\name{write_restart.SIBCASA}
\alias{write_restart.SIBCASA}
\title{Write restart template for SDA}
\usage{
write_restart.SIBCASA(
  outdir,
  runid,
  start.time,
  stop.time,
  settings,
  new.state
)
}
\arguments{
\item{outdir}{Output directory}

\item{runid}{Run ID}

\item{start.time}{Time of current assimilation step}

\item{stop.time}{Time of next assimilation step}

\item{settings}{PEcAn settings object}

\item{new.state}{Analysis state matrix returned by \code{sda.enkf}}
}
\description{
Write restart files for model
}
\author{
Alexey Shiklomanov
}
